Fix issue where debugger wasn't showing its contents.
Since the debugger window hosts a HWND, it should be initialized when the DebuggerView is inserted into the hierarchy, such that it gets sized properly. Also the DebuggerView needed to propagate changes to its own bounds into the inner view.

Also fixes a DCHECK in NavigationController that wasn't handling the existence of the debugger.

B=1302810

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@197 0039d316-1c4b-4281-b951-d872f2087c98