Initial version of a chrome memory test.  This is not yet done; as it still causes intermittent chrome crashes.  But, checking in an intermediate step so that others can play with it too.
Basic concept is to have a pre-loaded replay cache with interesting data.  This
test then drives chrome through automation, simulating a sequence of pages in
each tab
in a way that a user might also do so; this allows us to 
get per-tab caching of data similar to what the user would see.

Includes a logged in gmail session, a few pages that exercise some javascript,
several languages, and pages
from many mainstream sites.

Lots more can be done with this test going forward; for now the goal is to just
measure and track memory.



git-svn-id: svn://svn.chromium.org/chrome/trunk/src@361 0039d316-1c4b-4281-b951-d872f2087c98