Quick and easy fix for vista taskbar auto hide issue
In vista we designate the main frame border as a part
of non-client area. This is similar to a popup window having resizing 
border. The default logic to maximize such a window, stretches the
borders outside of the screen and the client area occupies the entire
work area. As a result windows taskbar treats it as a 'full screen'
application and hides itself in an auto-hide mode.

An XP-like fix for this, that is to reduce client area vertically by 
one pixel, would most likely involve overriding WM_SYSCOMMAND of 
SC_MAXIMIZE. Hence, this is a cheap, although a bit ugly fix in 
WM_GETMINMAXINFO handling.


BUG=20


Review URL: http://codereview.chromium.org/9305

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@4825 0039d316-1c4b-4281-b951-d872f2087c98