This fixes another issue with PDF documents loaded via FastWebView (NPN_RequestRead). This would occur whenever Chrome was configured to use a proxy server. The NPN_RequestRead API sends out byte range requests to the server, which then responds with a HTTP Partial Response header (206) followed by a content type which specifies the boundary string which separates individual body parts.
We parse out the boundary string in our plugin implementation and then send out individual body parts to the plugin.

In this case the proxy server would respond to the byte range request and send back the response with a quoted boundary string. This is legit as per MIME specs. We did not handle this case and hence would look for the quoted boundary string in the data which did not exist.

Fix is to trim the leading and trailing quotes in the boundary string.

Added a unit test to handle this case.

Bug=http://code.google.com/p/chromium/issues/detail?id=4076
R=jam

Review URL: http://codereview.chromium.org/10236

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@5132 0039d316-1c4b-4281-b951-d872f2087c98