Linux: calculate fractional xheight
We need to push the calculation of the x-height into Skia. For one of
the layout tests, it assumes that 2.5ex of Ahem is exactly 32px.
Previously we calculated the x-height of Ahem to be 13px, so 2.5*13 was
33px.

Now we use the 26.6 fixed point number from freetype's hinter and turn
that directly into a floating point value (x-height is the only WebKit
metric which is a floating point value).

This lets us pass LayoutTests/css2.1/t1507-c526-font-sz-02-b-a.html
(although the Windows baseline is wrong)

Review URL: http://codereview.chromium.org/13112


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6344 0039d316-1c4b-4281-b951-d872f2087c98