fix box shadows
We need to apply the context transform to the shadow before
rendering it.  There's also a weird case where we have to
flip the y offset if we're not applying the transform.  I'm
not sure why this is necessary, but it allows us to pass
canvas-shadow tests.

BUG=5503

Review URL: http://codereview.chromium.org/14125

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7002 0039d316-1c4b-4281-b951-d872f2087c98