Fixed crash when clicking empty select element.
I believe this is not needed:

  if (windowHeight == 0)  
    windowHeight = min(getRowHeight(-1), kMaxHeight);

windowHeight is dependent on the number of items within the popup, if you have
no items within the popup, it returns 0. I don't understand why we need to do
"min(getRowHeight(-1), kMaxHeight)" when its 0, that doesn't make sense to me.
Since getRowHeight gets the height of that item@index, everyone knows there are
no item for index -1 (hence crash).

Patch by Mohamed Mansour
R=darin
BUG=4334 (http://crbug.com/4334)
TEST=<select></select>



git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7347 0039d316-1c4b-4281-b951-d872f2087c98