Adobe Reader 7 expects the load_manually flag to be set when being instantiated. This only causes an issue when we reinitialize the plugin when we receive a HTTP 200 response for a byte range request.
We were setting the load_manually flag to false as we would be handing off the data to the plugin. However Reader 7 puts up a message box in its NPP_New call indicating that the operation failed. It then returns an error which causes the renderer to crash as we dereference a NULL plugin delegate
pointer.

We also force an invalidate when the plugin is reinitialized as the page does not paint at times.

The other fix is to pass down the plugin mime type correctly to WebPluginImpl.

This fixes http://code.google.com/p/chromium/issues/detail?id=6318

Bug=6318



Review URL: http://codereview.chromium.org/18831

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@8765 0039d316-1c4b-4281-b951-d872f2087c98