Fix dragging a Chrome window which has a single tab by the tab itself.
A change in 9953 made a tab of a single chrome-window with a single tab part of the
client area so that double-clicks on it are not treated as non-client double-clicks
and that the tab context-menu appears instead of the system context menu (see issue
3787). That change also solved issue 2827 (which is not restructed to the close
button of the right-most tab).

Change 10929 reverted 9953 partially, keeping the sinle-tab of a single-window as
part of the client area but removed the double-click swallowing code that was needed
to fix 2827. In doing so, dragging a single-tabbed single chrome window by the tab
(not the tabstrip) no longer worked. This small changes only fixes that behavior. It
turns out that there was a use case where some users want double-clicks to be considered
as two single-clicks (see issue 8287) and swallowing the double-click breaks this.

Review URL: http://codereview.chromium.org/46033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11973 0039d316-1c4b-4281-b951-d872f2087c98