Make the font family and the font size used in dom UI localizable. This is a part 1 of the fix for bug 7319 and will be followed by part 2 for non-DOM UI.
For some Indian languages (Malayalam, Bengali and Telugu), we have to 
use a bigger size and a font family for that script on Windows. 
This is because Windows' stock fonts for those scripts are smaller
than fonts for other scripts at a given size.

I removed 'WEB' style in chrome_font.h because it's not used anywhere any more
after our switch to the html UI. In addition, IDS_WEB_FONT_FAMILY is 
recycled to localize the font family (or the list of font families) for html UI.

I also back-ported the support for setting 'style.fooBar' property to our copy
of Jstemplate (JstProcessor.prototype.jstValues_).


BUG=7319

Review URL: http://codereview.chromium.org/57025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13114 0039d316-1c4b-4281-b951-d872f2087c98