POSIX: Don't allow onunload handlers to hang a renderer forever.
(Reland of r15025 which was reverted in r15095. |exit| has been
changed to |_exit| to save running the onexit handlers while another
thread is still in V8 code.)

On POSIX one can install an unload handler which loops forever and
leave behind a renderer process which eats 100% CPU forever.

This is because the terminate signals (ViewMsg_ShouldClose and the
error from the IPC channel) are routed to the main message loop but
never processes (because that message loop is stuck in V8).

One could make the browser SIGKILL the renderers, but that leaves open
a large window where a browser failure (or a user, manually
terminating the browser because "it's stuck") will leave behind a
process eating all the CPU.

On Windows we don't have this issue because all the processes are in a
job so when the parent dies, all the children are killed too.

http://codereview.chromium.org/100222


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@15332 0039d316-1c4b-4281-b951-d872f2087c98