Be slightly more precise in how we propagate mixed content state.
Mixed content can't really propagate from one process to another, so we don't need to break the lock icon for every tab in the same origin when we see insecure content.  We need only break the lock icon for all the tabs in the same process.

R=jcampan
TEST=SSLUITest.TestMixedContentsTwoTabs

Review URL: http://codereview.chromium.org/115230

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@15858 0039d316-1c4b-4281-b951-d872f2087c98