The features request broke backward compatibility for code currently in the wild. This CL fixes that as follows.
If there is no 'o3d_features' argument passed to the plugin it assumes the code
using the plugin is old and turns on the features that broke backward compatibility.
Otherwise, the utility libraries always pass in a o3d_features argument from this
point on so it will work the new way (ie, certain features must be requested).
Also, the requested version is passed in as well. This way we can adjust the API
by version number later.

Unfortunately, as it is, neither the o3d_features argument is required nor
is the version number which means if someone is not using the o3djs libraries
it would be possible for them to make something that we would break later.

We could maybe fix that by requiring a version number and failing if one is
not passed in  although old code will still not give us a version number so I'm
really not sure how to handle that case.

Review URL: http://codereview.chromium.org/118050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17994 0039d316-1c4b-4281-b951-d872f2087c98