Fix download shelf z ordering during close.
Pretty much the same as what we do in the find bar. I think the flickering thing in the bug desc. was another result of <http://crbug.com/14786>, as it was only present in 3.0.190.2 and not ToT, and toggling the bookmark bar fixed it.

http://crbug.com/15427
TEST=see bug

Review URL: http://codereview.chromium.org/149054

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19401 0039d316-1c4b-4281-b951-d872f2087c98