Ensure proper paths when saving pages with no title.
When saving a page with no title, such as a text file, attempt to
use a sane value for the save name by retrieving the last component
of the URL (if one exists).

This prevents the case where a page http://www.foo.com/a/path/name.txt
is saved as file "http www" with extension type "foo.com a path name.txt".

This change also makes the SavePackage code a little more flexible
and simpler to test.

BUG=none
TEST=Covered by unittest.

Review URL: http://codereview.chromium.org/155266

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20432 0039d316-1c4b-4281-b951-d872f2087c98